refactor(architecture): split CLI and phase owners
