+++ src/frame_compare/orchestration/execution.py
@@ -29,6 +29,7 @@ from frame_compare.orchestration.execution_types import (
     PublishPhaseOutput,
     ReportPhaseOutput,
 )
+from frame_compare.orchestration.phase_alignment import run_align_phase
 from frame_compare.orchestration.phase_output_application import apply_phase_output
 from frame_compare.orchestration.phase_post_render import (
     run_confirm_slowpics_upload_phase,
@@ -37,14 +38,11 @@ from frame_compare.orchestration.phase_post_render import (
     run_publish_phase,
     run_report_phase,
 )
+from frame_compare.orchestration.phase_render import run_render_phase
 from frame_compare.orchestration.phase_selection import (
     run_analyze_phase,
     select_initial_frame_plan,
     selection_label_for_frame,
-)
-from frame_compare.orchestration.phase_tasks import (
-    run_align_phase,
-    run_render_phase,
 )
 from frame_compare.orchestration.phases import Phase
 from frame_compare.orchestration.types import (
